+++ Classes/ormapper/TCategoryBase.m
@@ -45,7 +45,7 @@
   @param pid Primary key of the record
   @return record
 */
-+ (TCategory *)find:(int)pid
++ (TCategory *)find:(NSInteger)pid
 {
     Database *db = [Database instance];
 
@@ -86,7 +86,7 @@
   @param cond Conditions (ORDER BY etc)
   @note If you specify WHERE conditions, you must start cond with "AND" keyword.
 */
-+ (TCategory*)find_by_sorder:(int)key cond:(NSString *)cond
++ (TCategory*)find_by_sorder:(NSInteger)key cond:(NSString *)cond
 {
     if (cond == nil) {
         cond = @"WHERE sorder = ? LIMIT 1";
@@ -98,7 +98,7 @@
     return [self find_first_stmt:stmt];
 }
 
-+ (TCategory*)find_by_sorder:(int)key
++ (TCategory*)find_by_sorder:(NSInteger)key
 {
     return [self find_by_sorder:key cond:nil];
 }
@@ -303,7 +303,7 @@
     [s appendString:@","];
     [s appendString:[self quoteSqlString:_name]];
     [s appendString:@","];
-    [s appendString:[self quoteSqlString:[NSString stringWithFormat:@"%d", _sorder]]];
+    [s appendString:[self quoteSqlString:[NSString stringWithFormat:@"%ld", (long)_sorder]]];
     [s appendString:@");"];
 }
